requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
python-dotenv==1.0.0
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3
requests==2.31.0
aiohttp==3.9.1
asyncio==3.4.3
redis==5.0.1
python-pptx==0.6.23
markdown==3.5.1
python-markdown-math==0.8
pymupdf==1.23.8
openpyxl==3.1.2
pandas==2.1.4
python-jose[cryptography]==3.3.0
gunicorn==21.2.0
boto3==1.34.0
Werkzeug==3.0.1
marshmallow==3.20.1
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2-binary==2.9.9
celery==5.3.4
eventlet==0.33.3